server/ml-api/requirements.txt
fastapi
uvicorn
python-multipart
pillow
numpy
tensorflow
torch
opencv-python
ultralytics
